Issue #157 - Updating components to move js to  and compile the files.
