+++ components/alert/src/alert.js
@@ -1,0 +1,2 @@
+// Include alert javascript.
+import 'bootstrap/js/dist/alert';
