+++ components/header/header.js
@@ -1,72 +1,1 @@
-(function (Drupal, once) {
-  Drupal.behaviors.mainMenu = {
-    attach: function (context) {
-      const desktopBreakpoint = 768;
-
-      function setupDropdownBehavior(navElement) {
-        navElement.querySelectorAll('.navbar-nav .dropdown').forEach(item => {
-          const link = item.querySelector('.nav-link.dropdown-toggle');
-          const arrow = item.querySelector('.dropdown-arrow svg');
-          const dropdownMenu = item.querySelector('.dropdown-menu');
-
-          if (link && arrow && dropdownMenu) {
-            item.addEventListener('mouseenter', () => {
-              if (window.innerWidth >= desktopBreakpoint) openDropdown(item);
-            });
-
-            item.addEventListener('mouseleave', () => {
-              if (window.innerWidth >= desktopBreakpoint) closeDropdown(item);
-            });
-
-            link.addEventListener('click', e => {
-              window.location.href = link.getAttribute('href');
-            });
-
-            arrow.addEventListener('click', e => {
-              e.preventDefault();
-              toggleDropdown(item);
-            });
-          }
-        });
-      }
-
-      function openDropdown(item) {
-        item.classList.add('show');
-        item.querySelector('.dropdown-menu').style.cssText = 'max-height: 1200px; opacity: 1;';
-        const arrow = item.querySelector('.dropdown-arrow svg');
-        arrow.style.transform = 'rotate(180deg)';
-        arrow.setAttribute('aria-expanded', 'true');
-      }
-
-      function closeDropdown(item) {
-        item.classList.remove('show');
-        item.querySelector('.dropdown-menu').style.cssText = 'max-height: 0; opacity: 0;';
-        const arrow = item.querySelector('.dropdown-arrow svg');
-        arrow.style.transform = 'rotate(0deg)';
-        arrow.setAttribute('aria-expanded', 'false');
-      }
-
-      function toggleDropdown(item) {
-        if(item.classList.contains('show')) {
-          closeDropdown(item);
-        } else {
-          openDropdown(item);
-        }
-      }
-
-      function closeAllDropdowns(navElement) {
-        navElement.querySelectorAll('.nav-item.show').forEach(closeDropdown);
-      }
-
-
-      once('psul-theme-mainMenuNav', '#mainMenuNav', context).forEach(() => {
-        document.querySelectorAll('#mainMenuNav').forEach(setupDropdownBehavior);
-      });
-
-      once('psul-theme-offcanvasMain', '#offcanvasMain', context).forEach(() => {
-        document.querySelectorAll('#offcanvasMain').forEach(setupDropdownBehavior);
-      });
-
-    }
-  };
-})(Drupal, once);
+!function(e,n){e.behaviors.mainMenu={attach:function(e){function t(e){e.querySelectorAll(".navbar-nav .dropdown").forEach((function(e){var n=e.querySelector(".nav-link.dropdown-toggle"),t=e.querySelector(".dropdown-arrow svg"),a=e.querySelector(".dropdown-menu");n&&t&&a&&(e.addEventListener("mouseenter",(function(){window.innerWidth>=768&&o(e)})),e.addEventListener("mouseleave",(function(){window.innerWidth>=768&&r(e)})),n.addEventListener("click",(function(e){window.location.href=n.getAttribute("href")})),t.addEventListener("click",(function(n){n.preventDefault(),function(e){e.classList.contains("show")?r(e):o(e)}(e)})))}))}function o(e){e.classList.add("show"),e.querySelector(".dropdown-menu").style.cssText="max-height: 1200px; opacity: 1;";var n=e.querySelector(".dropdown-arrow svg");n.style.transform="rotate(180deg)",n.setAttribute("aria-expanded","true")}function r(e){e.classList.remove("show"),e.querySelector(".dropdown-menu").style.cssText="max-height: 0; opacity: 0;";var n=e.querySelector(".dropdown-arrow svg");n.style.transform="rotate(0deg)",n.setAttribute("aria-expanded","false")}n("psul-theme-mainMenuNav","#mainMenuNav",e).forEach((function(){document.querySelectorAll("#mainMenuNav").forEach(t)})),n("psul-theme-offcanvasMain","#offcanvasMain",e).forEach((function(){document.querySelectorAll("#offcanvasMain").forEach(t)}))}}}(Drupal,once);
