+++ components/sidebar_menu/sidebar_menu.js
@@ -1,38 +1,1 @@
-(function(Drupal, once) {
-  Drupal.behaviors.sidebarMenu = {
-    attach: function (context) {
-
-      function setupDropdownBehavior(navElement) {
-        navElement.querySelectorAll('.navbar-nav .dropdown').forEach(item => {
-          const arrow = item.querySelector('.dropdown-arrow svg');
-          const dropdownMenu = item.querySelector('.dropdown-menu');
-          arrow.addEventListener('click', e => {
-            e.preventDefault();
-            toggleDropdown(item, dropdownMenu);
-          });
-        });
-      }
-      function openDropdown(item, dropdownMenu) {
-        const arrow = item.querySelector('.dropdown-arrow svg');
-        arrow.setAttribute('aria-expanded', 'true');
-        item.classList.add('show');
-      }
-      function closeDropdown(item, dropdownMenu) {
-        const arrow = item.querySelector('.dropdown-arrow svg');
-        arrow.setAttribute('aria-expanded', 'false');
-        item.classList.remove('show');
-      }
-      function toggleDropdown(item, dropdownMenu) {
-        if (item.classList.contains('show')) {
-          closeDropdown(item, dropdownMenu);
-        } else {
-          openDropdown(item, dropdownMenu);
-        }
-      }
-      once('psul-theme-first-sidebar-accordion-menu', '.sidebar-menu-parent', context).forEach((element) => {
-        setupDropdownBehavior(element);
-      });
-    }
-  };
-
-})(Drupal, once);
+!function(e,r){e.behaviors.sidebarMenu={attach:function(e){function n(e){e.querySelectorAll(".navbar-nav .dropdown").forEach((function(e){var r=e.querySelector(".dropdown-arrow svg");e.querySelector(".dropdown-menu"),r.addEventListener("click",(function(r){r.preventDefault(),function(e,r){e.classList.contains("show")?function(e,r){var n=e.querySelector(".dropdown-arrow svg");n.setAttribute("aria-expanded","false"),e.classList.remove("show")}(e):function(e,r){var n=e.querySelector(".dropdown-arrow svg");n.setAttribute("aria-expanded","true"),e.classList.add("show")}(e)}(e)}))}))}r("psul-theme-first-sidebar-accordion-menu",".sidebar-menu-parent",e).forEach((function(e){n(e)}))}}}(Drupal,once);
